Add screens and draft navigation
