+++ apps/rufferal/src/app/screens/Screens.tsx
@@ -1,5 +1,6 @@
 import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
+import { ScreenProps } from '@rufferal/types';
 import { observer } from 'mobx-react-lite';
 import {
   CatAvatarScreen,
@@ -12,23 +13,23 @@ import {
   DogPersonalityScreen,
   ManagePetsScreen,
   PetSplashScreen,
-  SearchScreen,
+  ProfileSplashScreen,
+  SearchDashboardScreen,
+  SearchFiltersScreen,
+  SearchResultsScreen
 } from '.';
 
-const Stack = createNativeStackNavigator();
+const Stack = createNativeStackNavigator<ScreenProps>();
 
 export const Screens = observer(() => {
   return (
     <NavigationContainer>
       <Stack.Navigator screenOptions={{ headerShown: false }}>
         {/* ⬇️⬇️⬇️ CURRENT DEVELOPMENT PAGE ⬇️⬇️⬇️ */}
-        <Stack.Screen name="Search" component={SearchScreen} />
+        {/* <Stack.Screen name="Search" component={SearchDashboardScreen} /> */}
         {/* ⬆️⬆️⬆️ CURRENT DEVELOPMENT PAGE ⬆️⬆️⬆️ */}
 
-        <Stack.Screen
-          name="Create Pets Onboarding"
-          component={PetSplashScreen}
-        />
+        <Stack.Screen name="Pet Splash" component={PetSplashScreen} />
         <Stack.Screen name="Manage Pets" component={ManagePetsScreen} />
         <Stack.Screen name="Cat Details" component={CatDetailsScreen} />
         <Stack.Screen name="Cat Avatar" component={CatAvatarScreen} />
@@ -38,7 +39,10 @@ export const Screens = observer(() => {
         <Stack.Screen name="Dog Avatar" component={DogAvatarScreen} />
         <Stack.Screen name="Dog Personality" component={DogPersonalityScreen} />
         <Stack.Screen name="Dog Careplan" component={DogCareplanScreen} />
-        {/* <Stack.Screen name="Search" component={SearchScreen} /> */}
+        <Stack.Screen name="Search" component={SearchDashboardScreen} />
+        <Stack.Screen name="Search Filters" component={SearchFiltersScreen} />
+        <Stack.Screen name="Search Results" component={SearchResultsScreen} />
+        <Stack.Screen name="Profile Splash" component={ProfileSplashScreen} />
       </Stack.Navigator>
     </NavigationContainer>
   );
