+++ apps/rufferal/src/app/screens/index.ts
@@ -7,5 +7,8 @@ export { DogCareplanScreen } from './pets/dogs/DogCareplanScreen';
 export { DogDetailsScreen } from './pets/dogs/DogDetailsScreen';
 export { DogPersonalityScreen } from './pets/dogs/DogPersonalityScreen';
 export { ManagePetsScreen } from './pets/ManagePetsScreen';
-export { PetSplashScreen } from './profile/create/PetSplashScreen';
-export { SearchScreen } from './search/SearchScreen';
+export { PetSplashScreen } from './pets/PetSplashScreen';
+export { ProfileSplashScreen } from './profile/ProfileSplashScreen';
+export { SearchDashboardScreen } from './search/SearchDashboardScreen';
+export { SearchFiltersScreen } from './search/SearchFiltersScreen';
+export { SearchResultsScreen } from './search/SearchResultsScreen';
