+++ apps/rufferal/src/app/screens/pets/PetSplashScreen.tsx
@@ -1,0 +1,6 @@
+import { PageNavigationProps } from '@rufferal/types';
+import { PetSplash } from '@rufferal/ui';
+
+export const PetSplashScreen = ({ navigation }: PageNavigationProps) => {
+  return <PetSplash navigation={navigation} />;
+};
