+++ apps/rufferal/src/app/screens/profile/ProfileSplashScreen.tsx
@@ -1,0 +1,7 @@
+import { PageNavigationProps } from '@rufferal/types';
+import { ProfileSplash } from '@rufferal/ui';
+
+export const ProfileSplashScreen
+  = ({ navigation }: PageNavigationProps) => {
+  return <ProfileSplash navigation={navigation} />;
+};
