--- apps/rufferal/src/app/screens/profile/create/PetSplashScreen.tsx
@@ -1,6 +1,0 @@
-import { PageNavigationProps } from '@rufferal/types';
-import { PetSplash } from '@rufferal/ui';
-
-export const PetSplashScreen = ({ navigation }: PageNavigationProps) => {
-  return <PetSplash navigation={navigation} />;
-};
