+++ apps/rufferal/src/app/screens/search/SearchDashboardScreen.tsx
@@ -1,0 +1,6 @@
+import { PageNavigationProps } from '@rufferal/types';
+import { SearchDashboard } from '@rufferal/ui';
+
+export const SearchDashboardScreen = ({ navigation }: PageNavigationProps) => {
+  return <SearchDashboard navigation={navigation} />;
+};
