+++ apps/rufferal/src/app/screens/search/SearchFiltersScreen.tsx
@@ -1,0 +1,6 @@
+import { PageNavigationProps } from '@rufferal/types';
+import { SearchFilters } from '@rufferal/ui';
+
+export const SearchFiltersScreen = ({ navigation }: PageNavigationProps) => {
+  return <SearchFilters navigation={navigation} />;
+};
