+++ apps/rufferal/src/app/screens/search/SearchResultsScreen.tsx
@@ -1,0 +1,6 @@
+import { PageNavigationProps } from '@rufferal/types';
+import { SearchResults } from '@rufferal/ui';
+
+export const SearchResultsScreen = ({ navigation }: PageNavigationProps) => {
+  return <SearchResults navigation={navigation} />;
+};
