--- apps/rufferal/src/app/screens/search/SearchScreen.tsx
@@ -1,6 +1,0 @@
-import { PageNavigationProps } from '@rufferal/types';
-import { Search } from '@rufferal/ui';
-
-export const SearchScreen = ({ navigation }: PageNavigationProps) => {
-  return <Search navigation={navigation} />;
-};
