+++ ui/src/components/pages/index.ts
@@ -20,5 +20,8 @@ export { DogCareplan } from './pets/create/dogs/dog-careplan/dog-careplan';
 export { DogDetails } from './pets/create/dogs/dog-details/dog-details';
 export { DogPersonality } from './pets/create/dogs/dog-personality/dog-personality';
 export { ManagePets } from './pets/manage-pets/manage-pets';
-export { PetSplash } from './profiles/create/pet-splash/pet-splash';
-export { Search } from './search/search/search';
+export { PetSplash } from './pets/pet-splash/pet-splash';
+export { ProfileSplash } from './profile/profile-splash/profile-splash';
+export { SearchDashboard } from './search/search-dashboard/search-dashboard';
+export { SearchFilters } from './search/search-filters/search-filters';
+export { SearchResults } from './search/search-results/search-results';
