+++ ui/src/components/pages/pets/pet-splash/pet-splash.spec.tsx
@@ -1,0 +1,10 @@
+import { render } from '@testing-library/react-native';
+
+import { PetSplash } from './pet-splash';
+
+describe('PetSplash', () => {
+  it('should render successfully', () => {
+    const { root } = render(<PetSplash />);
+    expect(root).toBeTruthy();
+  });
+});
