+++ ui/src/components/pages/profile/profile-splash/profile-splash.spec.tsx
@@ -1,0 +1,10 @@
+import { render } from '@testing-library/react-native';
+
+import { ProfileSplash } from './profile-splash';
+
+describe('ProfileSplash', () => {
+  it('should render successfully', () => {
+    const { root } = render(<ProfileSplash />);
+    expect(root).toBeTruthy();
+  });
+});
