--- ui/src/components/pages/profiles/create/pet-splash/pet-splash.spec.tsx
@@ -1,10 +1,0 @@
-import { render } from '@testing-library/react-native';
-
-import { PetSplash } from './pet-splash';
-
-describe('PetSplash', () => {
-  it('should render successfully', () => {
-    const { root } = render(<PetSplash />);
-    expect(root).toBeTruthy();
-  });
-});
