+++ ui/src/components/pages/search/search-dashboard/search-dashboard.spec.tsx
@@ -1,0 +1,10 @@
+import { render } from '@testing-library/react-native';
+
+import { SearchDashboard } from './search-dashboard';
+
+describe('SearchDashboard', () => {
+  it('should render successfully', () => {
+    const { root } = render(<SearchDashboard />);
+    expect(root).toBeTruthy();
+  });
+});
