+++ ui/src/components/pages/search/search-filters/search-filters.spec.tsx
@@ -1,0 +1,10 @@
+import { render } from '@testing-library/react-native';
+
+import { SearchFilters } from './search-filters';
+
+describe('SearchFilters', () => {
+  it('should render successfully', () => {
+    const { root } = render(<SearchFilters />);
+    expect(root).toBeTruthy();
+  });
+});
